refactor(tests): Move UI tests to components directory
